requirements.txt
fastapi==0.109.0
pg8000
python-dotenv
SQLAlchemy
psycopg2-binary
Flask
